Using callback to process functions
Instead of passing lists around and all that nonsense, I decided to take a step back and change the design a bit. I added a foreach_function method for the metadata class that invokes a callback for each function it finds. This function can handle any additionaly processing for the target binary. In addition to that, I may just plop all the transformation code into tthat callback...