transformer analyzes functions
Integrated DynamoRIO, the transformer now checks if a function is safe for modification (i.e. it just tracks the stack delta to make sure we have a complete function). On another note, I feel real dumb for using dwarf and not the symbol table to find function boundaries :^) Fortunately, I designed the metadata class nicely so that I can just create a new 'backend' for it that uses the symbol table. For now dwarf will do fine.