Tracking function prologues and epilogues
I have all the information I need to do transformations. The error handling is not ideal (got a bit crazy, and it will catch when something goes wrong, but it's not as descriptive as I'd like). I'm not going to fret too bad about it since DynamoRIO has assertions in their production code, which makes the whole program crash if something bad happens (i.e. there is no fighting that). In any case, it's good enough for now.